CodeChef December Cook-Off 2019 - CHNGIT
Problem Statement : https://www.codechef.com/problems/CHNGIT
Given a sequence of integers, we need to make each element same, and for that we can choose one index, and change the element present at that index to any one of the adjacent elements. Task is to find the minimum number of times operation needs to be performed.